refactor: Moved tax table to its own file
